[#49][ADD] Posisbility to remove user access from case modal info
