+++ source/app/blueprints/case/templates/case.html
@@ -95,10 +95,10 @@
                             <small id="content_last_saved_by" class="mr-3 mt-1"></small>
                             <span id="last_saved" class="badge mr-3 ml-2"></span>
                             <small id="content_last_sync"></small>
-                            <button type="button" class="btn btn-xs btn-icon btn-round btn-outline-default mr-3 ml-3" onclick="sync_editor();">
-                                <i class="fas fa-sync rotate"></i>
+                            <button type="button" class="btn btn-sm btn-outline-default mr-2 ml-3" onclick="sync_editor();">
+                                Refresh
                             </button>
-                            <button class="btn btn-sm icon-note mr-3" onclick="edit_case_summary();"></button>
+                            <button class="btn btn-sm mr-3" onclick="edit_case_summary();">Edit</button>
                         </div>
                     </div>
                 </div>
@@ -204,6 +204,7 @@
 <script src="/static/assets/js/plugin/showdown/showdown.min.js"></script>
 <script src="/static/assets/js/iris/crc32.js"></script>
 <script src="/static/assets/js/iris/case.js"></script>
+<script src="/static/assets/js/iris/manage.cases.common.js"></script>
 <script src="/static/assets/js/iris/case.summary.js"></script>
 
 <script>
