+++ source/app/blueprints/manage/templates/access_control/manage_access-control.html
@@ -116,6 +116,7 @@
     <script src="/static/assets/js/plugin/datatables/dataTables.contextualActions.min.js"></script>
 
     <script src="/static/assets/js/iris/manage.users.js"></script>
+    <script src="/static/assets/js/iris/manage.cases.common.js"></script>
     <script src="/static/assets/js/iris/manage.groups.js"></script>
     <script src="/static/assets/js/iris/common.js"></script>
