+++ source/app/blueprints/manage/templates/manage_cases.html
@@ -161,6 +161,7 @@
 
 <script src="/static/assets/js/plugin/select/bootstrap-select.min.js"></script>
 
+<script src="/static/assets/js/iris/manage.cases.common.js"></script>
 <script src="/static/assets/js/iris/manage.cases.js"></script>
 
 <script>
