+++ source/app/blueprints/manage/templates/manage_customers.html
@@ -22,9 +22,9 @@
                                     <div class="card-title">Customers management</div>
                                 </div>
                                 <div class="col">
-                                    <button class="btn btn-sm btn-icon btn-round btn-primary-success float-right ml-3" onclick="add_customer();"><i class="fas fa-plus-circle"></i></button>
-                                    <button type="button" class="btn btn-sm btn-icon btn-round btn-outline-black float-right" onclick="refresh_customer_table(true);">
-                                        <i class="fas fa-sync rotate"></i>
+                                    <button class="btn btn-sm float-right ml-3" onclick="add_customer();">Add customer</button>
+                                    <button type="button" class="btn btn-sm btn-outline-black float-right" onclick="refresh_customer_table(true);">
+                                        Refresh
                                     </button>
                                 </div>
                             </div>
