+++ source/app/datamgmt/manage/manage_users_db.py
@@ -26,6 +26,7 @@ from app.iris_engine.access_control.utils import ac_access_level_mask_from_val_l
 from app.iris_engine.access_control.utils import ac_access_level_to_list
 from app.iris_engine.access_control.utils import ac_auto_update_user_effective_access
 from app.iris_engine.access_control.utils import ac_get_detailed_effective_permissions_from_groups
+from app.iris_engine.access_control.utils import ac_remove_case_access_from_user
 from app.models import Cases
 from app.models.authorization import CaseAccessLevel
 from app.models.authorization import Group
@@ -326,6 +327,25 @@ def remove_cases_access_from_user(user_id, cases_list):
 
     ac_auto_update_user_effective_access(user_id)
     return True, 'Cases access removed'
+
+
+def remove_case_access_from_user(user_id, case_id):
+    if not user_id or type(user_id) is not int:
+        return False, 'Invalid user id'
+
+    if not case_id or type(case_id) is not int:
+        return False, "Invalid case id"
+
+    UserCaseAccess.query.filter(
+        and_(
+            UserCaseAccess.case_id == case_id,
+            UserCaseAccess.user_id == user_id
+        )).delete()
+
+    db.session.commit()
+
+    ac_remove_case_access_from_user(user_id, case_id)
+    return True, 'Case access removed'
 
 
 def get_user_details(user_id):
